requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
gunicorn

# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\ims2file.py: 4
tqdm

# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\model.py: 3,4
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\modules\encoder.py: 4,5
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\modules\multihead_attention.py: 9,10,11,12
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\modules\transformer_decoder.py: 12,13,14,15
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\modules\utils.py: 15,18
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\output.py: 4,5
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\metrics.py: 7,8,9,10
torch

# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\modules\encoder.py: 3
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\output.py: 11
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\ims2file.py: 10
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\tb_visualizer.py: 9
torchvision



# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\__init__.py: 1
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\routes.py: 1
Flask
werkzeug

# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\output.py: 13
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\ims2file.py: 7
Pillow


# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\output.py: 3
matplotlib

# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\output.py: 15
tensorflow


# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\model.py: 6
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\modules\encoder.py: 7
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\modules\transformer_decoder.py: 18
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\output.py: 6
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\ims2file.py: 6
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\metrics.py: 6
# F:\PROJECTS\Flask Projects\Recipe Generation from Food Image\Foodimg2Ing\utils\tb_visualizer.py: 3
numpy
typing-extensions





